+++ src/Env/File/Finder/Options/EnvFileFinderOptionsInterface.php
@@ -4,26 +4,25 @@ declare(strict_types=1);
 
 namespace LDL\Env\File\Finder\Options;
 
+use LDL\File\Collection\Contracts\DirectoryCollectionInterface;
 use LDL\File\Contracts\FileInterface;
 use LDL\Framework\Base\Contracts\ArrayFactoryInterface;
 use LDL\Framework\Base\Contracts\JsonFactoryInterface;
 use LDL\Framework\Base\Contracts\JsonFileFactoryInterface;
 use LDL\Framework\Base\Contracts\Type\ToArrayInterface;
+use LDL\Type\Collection\Interfaces\Type\StringCollectionInterface;
 
 interface EnvFileFinderOptionsInterface extends ArrayFactoryInterface, ToArrayInterface, \JsonSerializable, JsonFactoryInterface, JsonFileFactoryInterface
 {
-    public function getFiles(): array;
+    public function getFiles(): StringCollectionInterface;
 
-    public function getDirectories(): array;
+    public function getDirectories(): DirectoryCollectionInterface;
 
-    public function getExcludedDirectories(): array;
+    public function getExcludedDirectories(): StringCollectionInterface;
 
-    public function getExcludedFiles(): array;
+    public function getExcludedFiles(): StringCollectionInterface;
 
-    /**
-     * @return EnvFileFinderOptionsInterface
-     */
-    public function merge(EnvFileFinderOptionsInterface $options);
+    public function merge(EnvFileFinderOptionsInterface $options): EnvFileFinderOptionsInterface;
 
-    public function write(string $file): FileInterface;
+    public function write(string $file, bool $force): FileInterface;
 }